Reg form moved to Forms folder and working again
