+++ src/components/LandingPage/Forms/registration-form.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import { Field, reduxForm, focus, } from 'redux-form';
-import { registerUser, } from '../../actions/users';
-import { login, } from '../../actions/auth';
-import Input from './input';
-import { required, nonEmpty, matches, length, isTrimmed, } from '../../validators';
+import { registerUser, } from '../../../actions/users';
+import { login, } from '../../../actions/auth';
+import Input from '.././input';
+import { required, nonEmpty, matches, length, isTrimmed, } from '../../../validators';
 const passwordLength = length({ min: 10, max: 72, });
 const matchesPassword = matches('password');
