+++ src/components/LandingPage/registration-page.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { connect, } from 'react-redux';
 import { Link, Redirect, } from 'react-router-dom';
-import RegistrationForm from './registration-form';
+import RegistrationForm from './Forms/registration-form.js';
 
 export function RegistrationPage(props) {
   // If we are logged in (which happens automatically when registration
